+++ src/utstring.h
@@ -201,7 +201,7 @@ _UNUSED_ static void _utstring_BuildTableR(
 {
     long i, j;
 
-    i = P_NeedleLen - 1;
+    i = (long)P_NeedleLen - 1;
     j = i + 1;
     P_KMP_Table[i + 1] = j;
     while (i >= 0)
@@ -278,8 +278,8 @@ _UNUSED_ static long _utstring_findR(
     long V_FindPosition = -1;
 
     /* Search from right to left. */
-    j = (P_HaystackLen - 1);
-    i = (P_NeedleLen - 1);
+    j = ((long)P_HaystackLen - 1);
+    i = ((long)P_NeedleLen - 1);
     while ( (j >= 0) && (j >= i) )
     {
         while ( (i < (int)P_NeedleLen) && (P_Needle[i] != P_Haystack[j]) )
@@ -314,13 +314,13 @@ _UNUSED_ static long utstring_find(
 
     if (P_StartPosition < 0)
     {
-        V_StartPosition = s->i + P_StartPosition;
+        V_StartPosition = (long)s->i + P_StartPosition;
     }
     else
     {
         V_StartPosition = P_StartPosition;
     }
-    V_HaystackLen = s->i - V_StartPosition;
+    V_HaystackLen = (long)s->i - V_StartPosition;
     if ( (V_HaystackLen >= (long)P_NeedleLen) && (P_NeedleLen > 0) )
     {
         V_KMP_Table = (long *)malloc(sizeof(long) * (P_NeedleLen + 1));
@@ -360,7 +360,7 @@ _UNUSED_ static long utstring_findR(
 
     if (P_StartPosition < 0)
     {
-        V_StartPosition = s->i + P_StartPosition;
+        V_StartPosition = (long)s->i + P_StartPosition;
     }
     else
     {
